requirements.txt
ultralytics~=8.3.24
pandas~=2.2.3
PyYAML~=6.0.2
requests~=2.32.3
Django~=4.2.16
djangorestframework~=3.15.2
opencv-contrib-python~=4.10.0.84
